chore: Renamed Logger.mts to logger.service.mts.
